+++ app/common/layout/PublicLayout/PublicLayout.tsx
@@ -1,4 +1,4 @@
-import React, { memo, useEffect } from "react";
+import React, { memo } from "react";
 
 import { ReactNodeNoStrings } from "degen/dist/types/types";
 import { Box, Text, useTheme } from "degen";
@@ -7,14 +7,9 @@ import ExtendedSidebar from "../../../modules/ExtendedSidebar/ExtendedSidebar";
 import Sidebar from "@/app/modules/Sidebar";
 import styled from "styled-components";
 import { useGlobal } from "@/app/context/globalContext";
-import Dashboard from "@/app/modules/Dashboard";
-import { useQuery } from "react-query";
-import { UserType } from "@/app/types";
-import { toast } from "react-toastify";
-import ConnectPage from "./ConnectPage";
 
 type PublicLayoutProps = {
-  children?: ReactNodeNoStrings;
+  children: ReactNodeNoStrings;
 };
 
 const Container = styled(Box)<{ issidebarexpanded: boolean }>`
@@ -47,44 +42,11 @@ const MobileContainer = styled(Box)`
   }
 `;
 
-const getUser = async () => {
-  const res = await fetch(`${process.env.API_HOST}/user/me`, {
-    credentials: "include",
-  });
-  return await res.json();
-};
-
 function PublicLayout(props: PublicLayoutProps) {
   const { children } = props;
-  const { isSidebarExpanded, connectedUser, connectUser } = useGlobal();
+  const { isSidebarExpanded } = useGlobal();
 
   const { mode } = useTheme();
-
-  const { data: currentUser, refetch } = useQuery<UserType>(
-    "getMyUser",
-    getUser,
-    {
-      enabled: false,
-    }
-  );
-
-  useEffect(() => {
-    if (!connectedUser && currentUser?.id) connectUser(currentUser.id);
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [currentUser, connectedUser]);
-
-  useEffect(() => {
-    refetch()
-      .then((res) => {
-        const data = res.data;
-        if (data?.id) connectUser(data.id);
-      })
-      .catch((err) => {
-        console.log(err);
-        toast.error("Could not fetch user data");
-      });
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
 
   // const { connect, connectors, isConnected } = useConnect();
 
@@ -101,26 +63,20 @@ function PublicLayout(props: PublicLayoutProps) {
         backgroundColor={mode === "dark" ? "background" : "backgroundSecondary"}
         id="public-layout"
       >
-        {!connectedUser && !currentUser?.id ? (
-          <ConnectPage />
-        ) : (
-          <>
-            <Sidebar />
-            <AnimatePresence initial={false}>
-              {isSidebarExpanded && <ExtendedSidebar />}
-            </AnimatePresence>
-            <Box
-              display="flex"
-              flexDirection="column"
-              width="full"
-              overflow="hidden"
-            >
-              <Container issidebarexpanded={isSidebarExpanded}>
-                {children}
-              </Container>
-            </Box>
-          </>
-        )}
+        <Sidebar />
+        <AnimatePresence initial={false}>
+          {isSidebarExpanded && <ExtendedSidebar />}
+        </AnimatePresence>
+        <Box
+          display="flex"
+          flexDirection="column"
+          width="full"
+          overflow="hidden"
+        >
+          <Container issidebarexpanded={isSidebarExpanded}>
+            {children}
+          </Container>
+        </Box>
       </DesktopContainer>
       <MobileContainer
         backgroundColor={mode === "dark" ? "background" : "backgroundSecondary"}
