+++ app/modules/Circle/CircleOverview/FolderView/CreateFolderItem.tsx
@@ -1,28 +1,12 @@
-import {
-  Box,
-  IconPlusSmall,
-  Button,
-  Text,
-  useTheme,
-  IconUserGroup,
-  IconLightningBolt,
-} from "degen";
+import { Box, IconPlusSmall, Button } from "degen";
 import { useState } from "react";
-import { ProjectOutlined } from "@ant-design/icons";
-import { Tooltip } from "react-tippy";
 
-interface Props {
-  setProjectModal: (a: boolean) => void;
-  setWorkstreamModal: (a: boolean) => void;
-  setRetroOpen: (a: boolean) => void;
-}
+import CreateSpaceModal from "../../CreateSpaceModal";
+import CreateProjectModal from "../../CreateProjectModal";
+import CreateRetro from "@/app/modules/Retro/CreateRetro";
 
-function CreateFolderItem({
-  setProjectModal,
-  setWorkstreamModal,
-  setRetroOpen,
-}: Props) {
-  const { mode } = useTheme();
+function CreateFolderItem({ folderId }: { folderId: string }) {
+
   const [hover, setHover] = useState(false);
 
   return (
@@ -47,48 +31,9 @@ function CreateFolderItem({
             gap={"3"}
             marginLeft="3"
           >
-            <Tooltip html={<Text>Create Project</Text>} theme={mode}>
-              <Button
-                size="small"
-                variant="transparent"
-                shape="circle"
-                onClick={(e) => {
-                  e.stopPropagation();
-                  setProjectModal(true);
-                }}
-              >
-                <ProjectOutlined
-                  style={{ fontSize: "1.1rem", color: "rgb(191, 90, 242, 1)" }}
-                />
-              </Button>
-            </Tooltip>
-            <Tooltip html={<Text>Create Workstream</Text>} theme={mode}>
-              <Button
-                data-tour="folder-create-workstream-button"
-                size="small"
-                variant="transparent"
-                shape="circle"
-                onClick={(e) => {
-                  e.stopPropagation();
-                  setWorkstreamModal(true);
-                }}
-              >
-                <IconUserGroup size={"4"} color="accent" />
-              </Button>
-            </Tooltip>
-            <Tooltip html={<Text>Create Retro</Text>} theme={mode}>
-              <Button
-                size="small"
-                variant="transparent"
-                shape="circle"
-                onClick={(e) => {
-                  e.stopPropagation();
-                  setRetroOpen(true);
-                }}
-              >
-                <IconLightningBolt size={"4"} color="accent" />
-              </Button>
-            </Tooltip>
+            <CreateProjectModal folderId={folderId} />
+            <CreateSpaceModal folderId={folderId} />
+            <CreateRetro folderId={folderId} />
           </Box>
         )}
       </Box>
