+++ app/modules/Circle/CircleOverview/index.tsx
@@ -5,9 +5,11 @@ import {
   Box,
   Button,
   Stack,
+  Text,
   useTheme,
   Input,
   IconSearch,
+  IconCollection,
   IconGrid,
 } from "degen";
 import { AnimatePresence } from "framer-motion";
@@ -23,8 +25,6 @@ import InviteMemberModal from "../ContributorsModal/InviteMembersModal";
 import { TypeView } from "./TypeView";
 import { FolderView } from "./FolderView";
 import Breadcrumbs from "@/app/common/components/Breadcrumbs";
-import Roles from "../RolesTab";
-import { FolderOpenOutlined } from "@ant-design/icons";
 
 interface Props {
   toggle: string;
@@ -57,7 +57,7 @@ const Toggle: FunctionComponent<Props> = ({ toggle, setToggle }) => {
         display: "block",
         padding: "0.2rem",
         borderRadius: "2rem",
-        width: "20rem",
+        width: "13.6rem",
         margin: "0rem auto",
         marginBottom: "0.5rem",
         boxShadow: `0px 1px 5px ${
@@ -75,21 +75,14 @@ const Toggle: FunctionComponent<Props> = ({ toggle, setToggle }) => {
         onClick={() => setToggle("Members")}
         bgcolor={toggle == "Members" ? true : false}
       >
-        Contributors
-      </ToggleButton>
-      <ToggleButton
-        onClick={() => setToggle("Roles")}
-        bgcolor={toggle == "Roles" ? true : false}
-      >
-        Roles
+        Members
       </ToggleButton>
     </Box>
   );
 };
 
 export default function CircleOverview() {
-  const { isSidebarExpanded, groupBy, setGroupBy, toggle, setToggle } =
-    useGlobal();
+  const { isSidebarExpanded, groupBy, setGroupBy } = useGlobal();
   const router = useRouter();
   const { circle: cId, retroSlug } = router.query;
   const {
@@ -97,10 +90,13 @@ export default function CircleOverview() {
     setIsBatchPayOpen,
     isBatchPayOpen,
     retro,
+    loading,
     navigationBreadcrumbs,
   } = useCircle();
   const { canDo } = useRoleGate();
   const [isRetroOpen, setIsRetroOpen] = useState(false);
+  const [toggle, setToggle] =
+    useState<"Overview" | "Members" | "Roles">("Overview");
   const [filteredProjects, setFilteredProjects] = useState(circle?.projects);
   const [filteredCollections, setFilteredCollections] = useState(
     circle?.collections
@@ -121,6 +117,26 @@ export default function CircleOverview() {
       setIsRetroOpen(true);
     }
   }, [retroSlug]);
+
+  if (circle?.unauthorized)
+    return (
+      <>
+        <Text size="headingTwo" weight="semiBold" ellipsis>
+          This circle is private
+        </Text>
+        <Button
+          size="large"
+          variant="transparent"
+          onClick={() => router.back()}
+        >
+          <Text size="extraLarge">Go Back</Text>
+        </Button>
+      </>
+    );
+
+  if (loading) {
+    return null;
+  }
 
   return (
     <>
@@ -163,7 +179,7 @@ export default function CircleOverview() {
               >
                 <Input
                   label=""
-                  placeholder="Search projects, workstreams, retros"
+                  placeholder="Search anything .."
                   prefix={<IconSearch />}
                   onChange={(e) => {
                     const proj = matchSorter(
@@ -217,11 +233,11 @@ export default function CircleOverview() {
                   {groupBy == "Type" ? (
                     <IconGrid size={"4"} />
                   ) : (
-                    <FolderOpenOutlined style={{ fontSize: "1.1rem" }}/>
+                    <IconCollection size={"4"} />
                   )}
                 </Button>
                 {canDo("inviteMembers") && (
-                  <Box width={"1/3"} marginBottom="1">
+                  <Box width={"1/3"} marginTop="2">
                     <InviteMemberModal />
                   </Box>
                 )}
@@ -247,7 +263,6 @@ export default function CircleOverview() {
             </Stack>
           )}
           {toggle == "Members" && <CircleMembers />}
-          {toggle == "Roles" && <Roles />}
         </Box>
       </Stack>
     </>
