+++ app/modules/Circle/CircleSettingsModal/GuildIntegration/index.tsx
@@ -44,7 +44,7 @@ export default function GuildIntegration() {
                         {
                           guildxyzId: guildServer.id,
                         },
-                        circle?.id
+                        circle?.id as string
                       );
                       console.log({ res });
                       setLoading(false);
