+++ app/modules/Circle/CircleSettingsModal/QuestbookIntegration/index.tsx
@@ -50,7 +50,7 @@ export default function ConnectQuestbook() {
         grantApplicantProject: applicantProject?.value,
         grantNotificationChannel: discordGrantNotifChannel,
       },
-      circle?.id
+      circle?.id as string
     );
     console.log({ res });
     setIsLoading(false);
@@ -85,7 +85,7 @@ export default function ConnectQuestbook() {
 
   useEffect(() => {
     const getGuildChannels = async () => {
-      const res = await fetchGuildChannels(circle?.discordGuildId);
+      const res = await fetchGuildChannels(circle?.discordGuildId as string);
       setChannels(
         res.guildChannels.map((channel: any) => ({
           label: channel.name,
