+++ app/modules/Circle/CircleSettingsModal/index.tsx
@@ -52,7 +52,7 @@ export default function SettingsModal({ handleClose, initialTab }: Props) {
         avatar: logo,
         private: visibilityTab === 1,
       },
-      circle?.id
+      circle?.id as string
     );
     setIsLoading(false);
     if (res) {
@@ -62,7 +62,7 @@ export default function SettingsModal({ handleClose, initialTab }: Props) {
   };
 
   const onDelete = async () => {
-    const res = await deleteCircle(circle?.id);
+    const res = await deleteCircle(circle?.id as string);
     if (res) {
       handleClose();
       void router.push("/");
