+++ app/modules/Circle/ContributorsModal/AddRoleModal/index.tsx
@@ -3,10 +3,8 @@ import Modal from "@/app/common/components/Modal";
 import PrimaryButton from "@/app/common/components/PrimaryButton";
 import CheckBox from "@/app/common/components/Table/Checkbox";
 import { addRole, updateRole } from "@/app/services/CircleRoles";
-import { EyeOutlined } from "@ant-design/icons";
 import {
   Box,
-  Button,
   IconCheck,
   IconPencil,
   IconPlusSmall,
@@ -18,7 +16,6 @@ import { AnimatePresence } from "framer-motion";
 import React, { useEffect, useState } from "react";
 import { useCircle } from "../../CircleContext";
 import { defaultPermissions } from "./contants";
-import useRoleGate from "@/app/services/RoleGate/useRoleGate";
 
 type props = {
   role?: string;
@@ -29,13 +26,12 @@ export default function AddRole({ role }: props) {
   const [isOpen, setIsOpen] = useState(false);
   const [permissions, setPermissions] = useState(defaultPermissions);
   const [loading, setLoading] = useState(false);
-  const { circle, fetchCircle } = useCircle();
-  const { canDo } = useRoleGate();
+  const { circle, setCircleData } = useCircle();
 
   useEffect(() => {
     if (role && circle) {
-      setName(circle.roles[role]?.name);
-      setPermissions(circle.roles[role]?.permissions);
+      setName(role);
+      setPermissions(circle.roles[role].permissions);
     }
   }, [circle, role]);
 
@@ -50,158 +46,112 @@ export default function AddRole({ role }: props) {
           Add Role
         </PrimaryButton>
       ) : (
-        <Button
-          prefix={
-            circle.roles[role]?.mutable && canDo("manageRoles") ? (
-              <IconPencil />
-            ) : (
-              <EyeOutlined />
-            )
-          }
+        <PrimaryButton
+          icon={<IconPencil />}
           onClick={() => setIsOpen(true)}
-          variant="transparent"
-          width={"full"}
-          size="small"
+          variant="tertiary"
         >
-          {circle.roles[role]?.mutable && canDo("manageRoles")
-            ? "Edit"
-            : "View"}{" "}
-          Permissions
-        </Button>
+          Edit
+        </PrimaryButton>
       )}
 
       <AnimatePresence>
         {isOpen && (
-          <Modal
-            title={
-              role
-                ? circle.roles[role]?.mutable && canDo("manageRoles")
-                  ? "Edit Permissions"
-                  : "View Permissions"
-                : "Add Role"
-            }
-            handleClose={() => setIsOpen(false)}
-          >
+          <Modal title="Add Role" handleClose={() => setIsOpen(false)}>
             <Box padding="8" overflow="auto" style={{ height: "40rem" }}>
               <Stack space="2">
-                {(!role || (role && circle.roles[role]?.mutable)) && (
-                  <>
-                    <Box marginLeft="4">
-                      <Text weight="semiBold">Role Name</Text>
-                    </Box>
-                    <Input
-                      label=""
-                      placeholder="guild_lead"
-                      value={name}
-                      onChange={(e) => setName(e.target.value)}
-                    />
-                  </>
-                )}
-                {role && !circle.roles[role]?.mutable && (
-                  <Box marginLeft="4">
-                    <Text weight="semiBold" align={"center"}>
-                      This role is immutable
-                    </Text>
-                  </Box>
-                )}
-
+                <Box marginLeft="4">
+                  <Text weight="semiBold">Role Name</Text>
+                </Box>
+                <Input
+                  label=""
+                  placeholder="guild_lead"
+                  value={name}
+                  onChange={(e) => setName(e.target.value)}
+                />
                 <Box marginLeft="4">
                   <Text weight="semiBold">Permissions</Text>
                 </Box>
                 {/* <Stack direction="horizontal" wrap> */}
-                {permissions &&
-                  Object.keys(permissions)?.map((key) => (
-                    <Box key={key} paddingY="2">
-                      <Stack direction="horizontal">
-                        <Box width="1/2">
-                          <Text variant="label">{key}</Text>
-                        </Box>
-                        {typeof (permissions as any)[key] === "boolean" ? (
-                          <CheckBox
-                            isChecked={(permissions as any)[key] as boolean}
-                            onClick={() => {
-                              setPermissions({
-                                ...permissions,
-                                [key]: !(permissions as any)[key],
-                              });
-                            }}
-                          />
-                        ) : (
-                          <Stack direction="horizontal">
-                            {Object.keys((permissions as any)[key]).map(
-                              (subKey) => (
-                                <Stack direction="horizontal" key={subKey}>
-                                  <Text>{subKey}</Text>
-                                  <CheckBox
-                                    // @ts-ignore
-                                    isChecked={permissions[key][subKey]}
-                                    onClick={() => {
-                                      setPermissions({
-                                        ...permissions,
-                                        [key]: {
-                                          // @ts-ignore
-                                          ...permissions[key],
-                                          // @ts-ignore
-                                          [subKey]: !permissions[key][subKey],
-                                        },
-                                      });
-                                    }}
-                                  />
-                                </Stack>
-                              )
-                            )}
-                          </Stack>
-                        )}
-                      </Stack>
-                    </Box>
-                  ))}
+                {Object.keys(permissions).map((key) => (
+                  <Box key={key} paddingY="2">
+                    <Stack direction="horizontal">
+                      <Box width="1/2">
+                        <Text variant="label">{key}</Text>
+                      </Box>
+                      {typeof (permissions as any)[key] === "boolean" ? (
+                        <CheckBox
+                          isChecked={(permissions as any)[key] as boolean}
+                          onClick={() => {
+                            setPermissions({
+                              ...permissions,
+                              [key]: !(permissions as any)[key],
+                            });
+                          }}
+                        />
+                      ) : (
+                        <Stack direction="horizontal">
+                          {Object.keys((permissions as any)[key]).map(
+                            (subKey) => (
+                              <Stack direction="horizontal" key={subKey}>
+                                <Text>{subKey}</Text>
+                                <CheckBox
+                                  // @ts-ignore
+                                  isChecked={permissions[key][subKey]}
+                                  onClick={() => {
+                                    setPermissions({
+                                      ...permissions,
+                                      [key]: {
+                                        // @ts-ignore
+                                        ...permissions[key],
+                                        // @ts-ignore
+                                        [subKey]: !permissions[key][subKey],
+                                      },
+                                    });
+                                  }}
+                                />
+                              </Stack>
+                            )
+                          )}
+                        </Stack>
+                      )}
+                    </Stack>
+                  </Box>
+                ))}
                 {/* </Stack> */}
-                {(!role ||
-                  (role &&
-                    circle.roles[role]?.mutable &&
-                    canDo("manageRoles"))) && (
-                  <PrimaryButton
-                    icon={<IconCheck />}
-                    loading={loading}
-                    onClick={async () => {
-                      setLoading(true);
-                      let res;
-                      if (role) {
-                        const payload = {
-                          name: name,
-                          description: role
-                            ? circle.roles[role]?.description
-                            : `${name} role`,
-                          selfAssignable: false,
-                          permissions,
-                        };
-                        console.log({ payload });
-                        res = await updateRole(circle?.id, role, payload);
-                      } else {
-                        const payload = {
-                          name: name,
-                          description: role
-                            ? circle.roles[role]?.description
-                            : `${name} role`,
-                          selfAssignable: false,
-                          permissions,
-                        };
-                        console.log({ payload });
-                        res = await addRole(circle?.id, payload);
-                      }
-                      console.log({ res });
-                      fetchCircle();
-                      setLoading(false);
-
-                      // reset
-                      setName("");
-                      setPermissions(defaultPermissions);
-                      res && setIsOpen(false);
-                    }}
-                  >
-                    Save
-                  </PrimaryButton>
-                )}
+                <PrimaryButton
+                  icon={<IconCheck />}
+                  loading={loading}
+                  onClick={async () => {
+                    setLoading(true);
+                    const payload = {
+                      role: name,
+                      description: name,
+                      selfAssignable: false,
+                      permissions,
+                    };
+                    console.log({ payload });
+                    let res;
+                    if (role) {
+                      res = await updateRole(
+                        circle?.id as string,
+                        role,
+                        payload
+                      );
+                    } else {
+                      res = await addRole(circle?.id as string, payload);
+                    }
+                    console.log({ res });
+                    setLoading(false);
+                    setCircleData(res);
+                    // reset
+                    setName("");
+                    setPermissions(defaultPermissions);
+                    res && setIsOpen(false);
+                  }}
+                >
+                  Save
+                </PrimaryButton>
               </Stack>
             </Box>
           </Modal>
