+++ app/modules/Circle/CreateSpaceModal.tsx
@@ -1,10 +1,16 @@
 import {
   Box,
+  Button,
+  IconPlusSmall,
   Input,
   MediaPicker,
   Stack,
+  IconUserGroup,
+  useTheme,
+  Text,
 } from "degen";
 import React, { useState } from "react";
+import { AnimatePresence } from "framer-motion";
 import { useRouter } from "next/router";
 import Loader from "@/app/common/components/Loader";
 import Modal from "@/app/common/components/Modal";
@@ -14,6 +20,7 @@ import PrimaryButton from "@/app/common/components/PrimaryButton";
 import { storeImage } from "@/app/common/utils/ipfs";
 import { useCircle } from "./CircleContext";
 import { updateFolder } from "@/app/services/Folders";
+import { Tooltip } from "react-tippy";
 
 type CreateWorkspaceDto = {
   name: string;
@@ -23,16 +30,12 @@ type CreateWorkspaceDto = {
   avatar: string;
 };
 
-interface Props {
-  setWorkstreamModal: (a: boolean) => void;
-  folderId?: string;
-}
-
-function CreateSpaceModal({ setWorkstreamModal, folderId }: Props) {
-
+function CreateSpaceModal({ folderId }: { folderId?: string }) {
+  const { mode } = useTheme();
+  const [modalOpen, setModalOpen] = useState(false);
   const [visibilityTab, setVisibilityTab] = useState(0);
   const onVisibilityTabClick = (id: number) => setVisibilityTab(id);
-  const close = () => setWorkstreamModal(false);
+  const close = () => setModalOpen(false);
 
   const [name, setName] = useState("");
   const [description, setDescription] = useState("");
@@ -112,48 +115,85 @@ function CreateSpaceModal({ setWorkstreamModal, folderId }: Props) {
   return (
     <>
       <Loader loading={isLoading} text="Creating your workstream" />
-      <Modal handleClose={close} title="Create Workstream">
-        <Box width="full" padding="8">
-          <Stack>
-            <Input
-              label=""
-              placeholder="Name"
-              value={name}
-              onChange={(e) => setName(e.target.value)}
-            />
-            <Input
-              label=""
-              placeholder="Description"
-              value={description}
-              onChange={(e) => setDescription(e.target.value)}
-            />
-            <MediaPicker
-              compact
-              defaultValue={{
-                type: "image/png",
-                url: logo,
-              }}
-              label="Choose or drag and drop media"
-              uploaded={!!logo}
-              onChange={uploadFile}
-              uploading={uploading}
-              maxSize={10}
-            />
-            <Tabs
-              selectedTab={visibilityTab}
-              onTabClick={onVisibilityTabClick}
-              tabs={["Public", "Private"]}
-              orientation="horizontal"
-              unselectedColor="transparent"
-            />
-            <Box width="full" marginTop="4">
-              <PrimaryButton onClick={onSubmit} disabled={name.length === 0}>
-                Create Workstream
-              </PrimaryButton>
+      {folderId ? (
+        <Tooltip html={<Text>Create Workstream</Text>} theme={mode}>
+          <Button
+            data-tour="folder-create-workstream-button"
+            size="small"
+            variant="transparent"
+            shape="circle"
+            onClick={(e) => {
+              e.stopPropagation();
+              setModalOpen(true);
+            }}
+          >
+            <IconUserGroup size={"4"} color="accent" />
+          </Button>
+        </Tooltip>
+      ) : (
+        <Button
+          data-tour="circle-create-workstream-button"
+          size="small"
+          variant="transparent"
+          shape="circle"
+          onClick={(e) => {
+            e.stopPropagation();
+            setModalOpen(true);
+          }}
+        >
+          <IconPlusSmall />
+        </Button>
+      )}
+
+      <AnimatePresence>
+        {modalOpen && (
+          <Modal handleClose={close} title="Create Workstream">
+            <Box width="full" padding="8">
+              <Stack>
+                <Input
+                  label=""
+                  placeholder="Name"
+                  value={name}
+                  onChange={(e) => setName(e.target.value)}
+                />
+                <Input
+                  label=""
+                  placeholder="Description"
+                  value={description}
+                  onChange={(e) => setDescription(e.target.value)}
+                />
+                <MediaPicker
+                  compact
+                  defaultValue={{
+                    type: "image/png",
+                    url: logo,
+                  }}
+                  label="Choose or drag and drop media"
+                  uploaded={!!logo}
+                  onChange={uploadFile}
+                  uploading={uploading}
+                  maxSize={10}
+                />
+                <Tabs
+                  selectedTab={visibilityTab}
+                  onTabClick={onVisibilityTabClick}
+                  tabs={["Public", "Private"]}
+                  orientation="horizontal"
+                  unselectedColor="transparent"
+                />
+                <Box width="full" marginTop="4">
+                  <PrimaryButton
+                    onClick={onSubmit}
+                    disabled={name.length === 0}
+                  >
+                    Create Workstream
+                  </PrimaryButton>
+                </Box>
+              </Stack>
             </Box>
-          </Stack>
-        </Box>
-      </Modal>
+          </Modal>
+        )}
+      </AnimatePresence>
     </>
   );
 }
