+++ app/modules/Circle/index.tsx
@@ -2,7 +2,7 @@ import Loader from "@/app/common/components/Loader";
 import useCircleOnboarding from "@/app/services/Onboarding/useCircleOnboarding";
 import useRoleGate from "@/app/services/RoleGate/useRoleGate";
 import { ApartmentOutlined } from "@ant-design/icons";
-import { Box, Button, useTheme, Text, IconLockClosed, Stack } from "degen";
+import { Box, Button, useTheme, Text } from "degen";
 import { AnimatePresence } from "framer-motion";
 import { useState } from "react";
 import { ToastContainer } from "react-toastify";
@@ -13,23 +13,15 @@ import { useCircle } from "./CircleContext";
 import Onboarding from "./CircleOnboarding";
 import CircleOverview from "./CircleOverview";
 import { useRouter } from "next/router";
-import { joinCircle } from "@/app/services/JoinCircle";
-import Roles from "./RolesTab";
+import { useGlobal } from "@/app/context/globalContext";
 
 const BoxContainer = styled(Box)`
   width: calc(100vw - 4rem);
 `;
 
 export default function Circle() {
-  const {
-    circle,
-    isLoading,
-    memberDetails,
-    page,
-    fetchCircle,
-    fetchMemberDetails,
-  } = useCircle();
-
+  const { circle, isLoading, memberDetails, page } = useCircle();
+  const { connectedUser } = useGlobal();
   const { canDo } = useRoleGate();
   const { onboarded } = useCircleOnboarding();
   const { mode } = useTheme();
@@ -40,52 +32,20 @@ export default function Circle() {
     return <Loader text="...." loading />;
   }
 
-  if (circle?.unauthorized && !isLoading && circle?.id)
+  if (circle?.unauthorized)
     return (
-      <Box marginX="6">
-        <Box
-          style={{
-            margin: "1rem auto",
-            display: "flex",
-            flexDirection: "column",
-            alignItems: "center",
-            gap: "1rem",
-          }}
+      <BoxContainer padding={"16"}>
+        <Text size="headingTwo" weight="semiBold" ellipsis>
+          This circle is private
+        </Text>
+        <Button
+          size="large"
+          variant="transparent"
+          onClick={() => router.back()}
         >
-          <Button shape="circle" variant="secondary" size="large">
-            <IconLockClosed color={"accent"} size="6" />
-          </Button>
-
-          <Text size="headingTwo" weight="semiBold" ellipsis>
-            This circle is private
-          </Text>
-          <Stack direction="horizontal" space="4" align="center">
-            <Button
-              size="small"
-              variant="transparent"
-              onClick={() => router.back()}
-            >
-              <Text size="base">Go Back</Text>
-            </Button>
-            <Button
-              size="small"
-              variant="secondary"
-              onClick={async () => {
-                const data = await joinCircle(circle.id);
-                if (data) {
-                  fetchCircle();
-                  fetchMemberDetails();
-                }
-              }}
-            >
-              <Text size="base" color={"accent"}>
-                Get Role
-              </Text>
-            </Button>
-          </Stack>
-        </Box>
-        <Roles />
-      </Box>
+          <Text size="extraLarge">Go Back</Text>
+        </Button>
+      </BoxContainer>
     );
 
   return (
