+++ app/modules/ExtendedSidebar/ExtendedSidebar.tsx
@@ -12,11 +12,42 @@ import { useGlobal } from "@/app/context/globalContext";
 import ProfileButton from "../Sidebar/ProfileButton";
 import { Connect } from "../Sidebar/ProfileButton/ConnectButton";
 
+const getUser = async () => {
+  const res = await fetch(`${process.env.API_HOST}/user/me`, {
+    credentials: "include",
+  });
+  return await res.json();
+};
+
 function ExtendedSidebar(): ReactElement {
   const router = useRouter();
   const { circle: cId } = router.query;
-  const { connectedUser } = useGlobal();
+  const { connectUser, connectedUser } = useGlobal();
+  const { data: currentUser, refetch } = useQuery<UserType>(
+    "getMyUser",
+    getUser,
+    {
+      enabled: false,
+    }
+  );
 
+  useEffect(() => {
+    if (!connectedUser && currentUser?.id) connectUser(currentUser.id);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [currentUser, connectedUser]);
+
+  useEffect(() => {
+    refetch()
+      .then((res) => {
+        const data = res.data;
+        if (data?.id) connectUser(data.id);
+      })
+      .catch((err) => {
+        console.log(err);
+        toast.error("Could not fetch user data");
+      });
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
 
   return (
     <motion.div
