+++ app/modules/Profile/TaskWallet/TaskWalletTab/index.tsx
@@ -31,6 +31,7 @@ export const Card = styled(Box)<{ mode: string }>`
   flex-direction: column;
   flex-wrap: wrap;
   width: 630px;
+  // max-height: 300px;
   padding: 0.6rem;
   border-radius: 0.5rem;
   background-color: transparent;
@@ -126,7 +127,7 @@ const TaskWalletTabs = ({ userData, tab }: UserProps) => {
       </Box>
       <Box>
         {panelTab === "Work" && (
-          <Box display={"flex"} flexDirection="column">
+          <Box>
             <Toggle toggle={toggle} setToggle={setToggle} />
             <WorkCards
               toggle={toggle}
