+++ app/modules/Project/BoardView/index.tsx
@@ -112,7 +112,6 @@ function BoardView({ viewId }: Props) {
           advFilters?.groupBy == "Status" &&
           project?.columnOrder?.map((columnId, index): any => {
             const column = project.columnDetails[columnId];
-            if (!column) return null;
             let cards = column.cards?.map(
               (cardId: string) => filteredCards[cardId]
             );
