+++ app/services/JoinCircle/index.ts
@@ -96,10 +96,7 @@ export const joinCircle = async (circleId: string) => {
     console.log({ data });
     return data;
   } else {
-    toast("Sorry, you dont meet the requirements to join this circle", {
-      theme: "dark",
-    });
-    console.log("Error joining circle");
+    toast.error("Sorry, you dont meet the requirements to join this circle");
     return null;
   }
 };
