+++ app/types/index.d.ts
@@ -32,12 +32,6 @@ interface UserType {
   followedUsers: string[];
   followedByUsers: string[];
   userDetails: any;
-  circleDetails: {
-    [key: string]: {
-      slug: string;
-      avatar: string;
-    }
-  }
 }
 
 export interface Payment {
@@ -224,7 +218,7 @@ export interface CircleType {
     [key: string]: string[];
   };
   roles: {
-    [role: string]: {
+    [name: string]: {
       name: string;
       description: string;
       permissions: Permissions;
@@ -476,7 +470,6 @@ export interface CardDetails {
     id: string;
   }[];
   assignee: {
-    ethAddress: string | undefined;
     username: string;
     avatar: string;
     id: string;
