+++ public/sitemap-0.xml
@@ -1,7 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:news="http://www.google.com/schemas/sitemap-news/0.9" xmlns:xhtml="http://www.w3.org/1999/xhtml" xmlns:mobile="http://www.google.com/schemas/sitemap-mobile/1.0" xmlns:image="http://www.google.com/schemas/sitemap-image/1.1" xmlns:video="http://www.google.com/schemas/sitemap-video/1.1">
-<url><loc>https://circles.spect.network</loc><lastmod>2022-10-18T21:11:19.282Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority><xhtml:link rel="alternate" hreflang="en" href="https://circles.spect.network/en"/><xhtml:link rel="alternate" hreflang="de" href="https://circles.spect.network/de"/></url>
-<url><loc>https://circles.spect.network/explore</loc><lastmod>2022-10-18T21:11:19.282Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority><xhtml:link rel="alternate" hreflang="en" href="https://circles.spect.network/en/explore"/><xhtml:link rel="alternate" hreflang="de" href="https://circles.spect.network/de/explore"/></url>
-<url><loc>https://circles.spect.network/linkDiscord</loc><lastmod>2022-10-18T21:11:19.282Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority><xhtml:link rel="alternate" hreflang="en" href="https://circles.spect.network/en/linkDiscord"/><xhtml:link rel="alternate" hreflang="de" href="https://circles.spect.network/de/linkDiscord"/></url>
-<url><loc>https://circles.spect.network/linkGithub</loc><lastmod>2022-10-18T21:11:19.282Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority><xhtml:link rel="alternate" hreflang="en" href="https://circles.spect.network/en/linkGithub"/><xhtml:link rel="alternate" hreflang="de" href="https://circles.spect.network/de/linkGithub"/></url>
+<url><loc>https://circles.spect.network</loc><lastmod>2022-09-29T12:01:10.990Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority><xhtml:link rel="alternate" hreflang="en" href="https://circles.spect.network/en"/><xhtml:link rel="alternate" hreflang="de" href="https://circles.spect.network/de"/></url>
+<url><loc>https://circles.spect.network/linkDiscord</loc><lastmod>2022-09-29T12:01:10.990Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority><xhtml:link rel="alternate" hreflang="en" href="https://circles.spect.network/en/linkDiscord"/><xhtml:link rel="alternate" hreflang="de" href="https://circles.spect.network/de/linkDiscord"/></url>
+<url><loc>https://circles.spect.network/linkGithub</loc><lastmod>2022-09-29T12:01:10.990Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority><xhtml:link rel="alternate" hreflang="en" href="https://circles.spect.network/en/linkGithub"/><xhtml:link rel="alternate" hreflang="de" href="https://circles.spect.network/de/linkGithub"/></url>
 </urlset>
